+++ src/onegov.feriennet/onegov/feriennet/layout.py
@@ -1,5 +1,6 @@
 from cached_property import cached_property
 from onegov.activity import Activity, PeriodCollection, Occasion
+from onegov.activity import BookingCollection
 from onegov.feriennet import _
 from onegov.feriennet import security
 from onegov.feriennet.collections import BillingCollection
@@ -149,6 +150,29 @@ class BookingCollectionLayout(DefaultLayout):
             Link(_("Homepage"), self.homepage_url),
             Link(self.title, self.request.link(self.model))
         ]
+
+
+class GroupInviteLayout(DefaultLayout):
+
+    @cached_property
+    def breadcrumbs(self):
+        wishlist_phase = self.app.active_period \
+            and self.app.active_period.wishlist_phase
+
+        if self.request.is_logged_in:
+            return [
+                Link(_("Homepage"), self.homepage_url),
+                Link(
+                    wishlist_phase and _("Wishlist") or _("Bookings"),
+                    self.request.class_link(BookingCollection)
+                ),
+                Link(_("Group"), '#')
+            ]
+        else:
+            return [
+                Link(_("Homepage"), self.homepage_url),
+                Link(_("Group"), '#')
+            ]
 
 
 class VacationActivityFormLayout(DefaultLayout):